+++ components-sdk/src/main/java/co/xendit/components/ui/helper/CurrencyUtil.kt
@@ -1,10 +1,11 @@
 package co.xendit.components.ui.helper
 
+import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.Locale
 
 internal object CurrencyUtil {
-  fun formatAmount(amount: Long?, currency: String?): String {
+  fun formatAmount(amount: BigDecimal?, currency: String?): String {
     if (amount == null || currency.isNullOrBlank()) return ""
     val symbol =
       when (currency) {
@@ -13,7 +14,8 @@ internal object CurrencyUtil {
         "PHP" -> "₱"
         else -> currency
       }
-    val number = NumberFormat.getNumberInstance(Locale.Builder().setLanguage("id").setRegion("ID").build())
+    val number =
+      NumberFormat.getNumberInstance(Locale.Builder().setLanguage("id").setRegion("ID").build())
     return "$symbol${number.format(amount)}"
   }
 }
